Class-ifying the vis system.
Main class now vis::Control
There is however still a global instance of it (vis::controller)

The vis:: globals are now mostly instance variables of this.

The physical min/max pressure, velocity and stress are moved to lbm_phys_pressure_max etc globals, pending the refactoring of that.

--HG--
branch : 2010-refactor