Avoid repeated calls to MPI_Comm_size
When using net::MpiCommunicator, net.Size() is implemented by a call to MPI_Comm_size. Using it as the upper bound in this loop actually caused it to dominate the time spent in the RequestComms phase. I noticed this reviewing some old MAP files from the ARCHER work - at 8k cores this actually consumed 12% of wall-clock time during the RunSimulation() phase. At 16k cores it took 45% of the time. This was for very short iteration timesteps but considering the simplicity of the fix and the apparent O(N^2) in #procs it'll probably manifest in production scale runs sooner or later as we continue to reach for higher scales.